requirements.txt
aiogram==2.25.1
aiohttp==3.8.6

python-dotenv==1.0.1
pydantic==2.9.2

langchain==0.1.0
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf==3.17.4
langchain-community==0.0.10